backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.3.3
Pillow==11.3.0
django-filter==23.1